# docs/swagger.yaml
basePath: /
definitions:
...
  ds.Task:
    properties:
      description:
        type: string
      image:
        type: string
      minidescription:
        type: string
      name:
        type: string
      status:
        type: string
      subject:
        type: string
      task_id:
        type: integer
    type: object
...
host: 127.0.0.1
info:
  contact:
    email: bitop@spatecon.ru
    name: API Support
    url: https://vk.com/bmstu_schedule
  description: Bmstu Open IT Platform
  license:
    name: AS IS (NO WARRANTY)
  title: BITOP
  version: "1.0"
paths:
...
  /task-request/delete/task/{id_c}/request/{id_r}:
    delete:
      consumes:
      - application/json
      description: delete task from request
      operationId: delete-task-from-request
      parameters:
      - description: ID задания
        in: path
        name: id_c
        required: true
        type: integer
      - description: ID заявки
        in: path
        name: id_r
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Консультация была удалена из заявки
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            type: string
        "404":
          description: Некорректный запрос
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete Task From Request
      tags:
      - Task-Request
  /tasks:
    get:
      description: Get all tasks
      operationId: get-tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.Task'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Task'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Task'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Task'
      summary: Get Tasks
      tags:
      - Tasks
  /tasks/{id}:
    get:
      consumes:
      - application/json
      description: Show task by ID
      operationId: get-task-by-id
      parameters:
      - description: ID задания
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.Task'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Task'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Task'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Task'
      summary: Get Task by ID
      tags:
      - Tasks
  /tasks/{id}/add-to-request:
    post:
      consumes:
      - application/json
      description: add task to request
      operationId: add-task-to-request
      parameters:
      - description: ID задания
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Task'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Task'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Task'
      security:
      - ApiKeyAuth: []
      summary: add task to request
      tags:
      - Tasks
  /tasks/{id}/addImage:
    post:
      consumes:
      - application/json
      description: Add an image to a specific task by ID.
      operationId: add-task-image
      parameters:
      - description: ID задания
        in: path
        name: id
        required: true
        type: integer
      - description: Image file to be uploaded
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Task'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Task'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Task'
      security:
      - ApiKeyAuth: []
      summary: Add task image
      tags:
      - Tasks
  /tasks/create:
    post:
      consumes:
      - application/json
      description: create task
      operationId: create-task
      parameters:
      - description: task info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ds.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Task'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Task'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Task'
      security:
      - ApiKeyAuth: []
      summary: create task
      tags:
      - Tasks
  /tasks/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete task by ID
      operationId: delete-task-by-id
      parameters:
      - description: ID задания
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Task'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Task'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Task'
      security:
      - ApiKeyAuth: []
      summary: Delete task by ID
      tags:
      - Tasks
  /tasks/request/{id}:
    get:
      consumes:
      - application/json
      description: Show task by ID of request
      operationId: get-task-by-id-of-request
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.Task'
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Task'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Task'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Task'
      security:
      - ApiKeyAuth: []
      summary: Get Task by request ID
      tags:
      - Tasks
  /tasks/update/{id}:
    put:
      consumes:
      - application/json
      description: update task
      operationId: update-task
      parameters:
      - description: ID задания
        in: path
        name: id
        required: true
        type: integer
      - description: task info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/ds.Task'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Task'
        "404":
          description: Некорректный запрос
          schema:
            $ref: '#/definitions/ds.Task'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/ds.Task'
      security:
      - ApiKeyAuth: []
      summary: update task
      tags:
      - Tasks
schemes:
- http
swagger: "2.0"